Move parse_args to separate util
